Introduce CVM
